// examples/petstore/generated/types.ts
/**
 * -----------------------------------
 * types.ts
 * -----------------------------------
 * Types and Enums are generated into this file for usage in both the front and backend of your typescript application, representing data coming from and going to the api.
 *
 * This is a generated file, do not directly edit. It was generated from the
 * openapi spec defined in your project.
 *
 * Generated by framer@1.0.0
 */

export type NonEmptyArray<T> = [T, ...T[]];

/**
 * Type: Pet
 */
export interface Pet {
    id?: number;
    name: string;
    category: Category | null | undefined;
    photoUrls: NonEmptyArray<string>;
    tags: Tag[];
    status: StatusEnum | null | undefined;
}

/**
 * Type: Category
 */
export interface Category {
    id?: number;
    name?: string;
}

/**
 * Type: Tag
 */
export interface Tag {
    id?: number;
    name?: string;
}

/**
 * Enum: StatusEnum
 */
export enum StatusEnum {
    Placed = "placed",
    Approved = "approved",
    Delivered = "delivered",
}

/**
 * Type: Error
 */
export interface Error {
    code: string;
    message: string;
}

/**
 * Type: ApiResponse
 */
export interface ApiResponse {
    code?: number;
    type?: string;
    message?: string;
}

/**
 * Type: Order
 */
export interface Order {
    id?: number;
    petId?: number;
    quantity?: number;
    shipDate?: string;
    status: StatusEnum | null | undefined;
    complete?: boolean;
}

/**
 * Type: User
 */
export interface User {
    id?: number;
    username?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    password?: string;
    phone?: string;
    userStatus?: number;
}